+++ packages/rich-text-html-renderer/src/__test__/documents/index.ts
@@ -6,6 +6,7 @@ export { default as paragraphDoc } from './paragraph';
 export { default as headingDoc } from './heading';
 export { default as marksDoc } from './mark';
 export { default as embeddedEntryDoc } from './embedded-entry';
+export { default as embeddedResourceDoc } from './embedded-resource';
 export { default as olDoc } from './ol';
 export { default as ulDoc } from './ul';
 export { default as quoteDoc } from './quote';