+++ packages/rich-text-react-renderer/src/index.tsx
@@ -12,6 +12,7 @@ const defaultNodeRenderers: RenderNode = {
   [BLOCKS.HEADING_5]: (node, children) => <h5>{children}</h5>,
   [BLOCKS.HEADING_6]: (node, children) => <h6>{children}</h6>,
   [BLOCKS.EMBEDDED_ENTRY]: (node, children) => <div>{children}</div>,
+  [BLOCKS.EMBEDDED_RESOURCE]: (node, children) => <div>{children}</div>,
   [BLOCKS.UL_LIST]: (node, children) => <ul>{children}</ul>,
   [BLOCKS.OL_LIST]: (node, children) => <ol>{children}</ol>,
   [BLOCKS.LIST_ITEM]: (node, children) => <li>{children}</li>,
@@ -27,7 +28,11 @@ const defaultNodeRenderers: RenderNode = {
   [BLOCKS.TABLE_CELL]: (node, children) => <td>{children}</td>,
   [INLINES.ASSET_HYPERLINK]: (node) => defaultInline(INLINES.ASSET_HYPERLINK, node as Inline),
   [INLINES.ENTRY_HYPERLINK]: (node) => defaultInline(INLINES.ENTRY_HYPERLINK, node as Inline),
+  [INLINES.RESOURCE_HYPERLINK]: (node) =>
+    defaultInlineResource(INLINES.RESOURCE_HYPERLINK, node as Inline),
   [INLINES.EMBEDDED_ENTRY]: (node) => defaultInline(INLINES.EMBEDDED_ENTRY, node as Inline),
+  [INLINES.EMBEDDED_RESOURCE]: (node, children) =>
+    defaultInlineResource(INLINES.EMBEDDED_RESOURCE, node as Inline),
   [INLINES.HYPERLINK]: (node, children) => <a href={node.data.uri}>{children}</a>,
 };
 
@@ -38,12 +43,21 @@ const defaultMarkRenderers: RenderMark = {
   [MARKS.CODE]: (text) => <code>{text}</code>,
   [MARKS.SUPERSCRIPT]: (text) => <sup>{text}</sup>,
   [MARKS.SUBSCRIPT]: (text) => <sub>{text}</sub>,
+  [MARKS.STRIKETHROUGH]: (text) => <s>{text}</s>,
 };
 
 function defaultInline(type: string, node: Inline): ReactNode {
   return (
     <span key={node.data.target.sys.id}>
       type: {node.nodeType} id: {node.data.target.sys.id}
+    </span>
+  );
+}
+
+function defaultInlineResource(type: string, node: Inline) {
+  return (
+    <span key={node.data.target.sys.urn}>
+      type: {node.nodeType} urn: {node.data.target.sys.urn}
     </span>
   );
 }
@@ -79,6 +93,10 @@ export interface Options {
    * Text renderer
    */
   renderText?: RenderText;
+  /**
+   * Keep line breaks and multiple spaces
+   */
+  preserveWhitespace?: boolean;
 }
 
 /**
@@ -102,5 +120,6 @@ export function documentToReactComponents(
       ...options.renderMark,
     },
     renderText: options.renderText,
+    preserveWhitespace: options.preserveWhitespace,
   });
 }