+++ packages/rich-text-types/src/emptyDocument.ts
@@ -5,7 +5,7 @@ import { BLOCKS } from './blocks';
  * A rich text document considered to be empty.
  * Any other document structure than this is not considered empty.
  */
-const EMPTY_DOCUMENT: Document = {
+export const EMPTY_DOCUMENT: Document = {
   nodeType: BLOCKS.DOCUMENT,
   data: {},
   content: [
@@ -23,5 +23,3 @@ const EMPTY_DOCUMENT: Document = {
     },
   ],
 };
-
-export default EMPTY_DOCUMENT;