+++ packages/rich-text-types/src/marks.ts
@@ -1,13 +1,12 @@
 /**
  * Map of all Contentful marks.
  */
-enum MARKS {
+export enum MARKS {
   BOLD = 'bold',
   ITALIC = 'italic',
   UNDERLINE = 'underline',
   CODE = 'code',
   SUPERSCRIPT = 'superscript',
   SUBSCRIPT = 'subscript',
+  STRIKETHROUGH = 'strikethrough',
 }
-
-export default MARKS;