+++ packages/rich-text-types/tools/jsonSchemaGen.ts
@@ -1,8 +1,8 @@
-import { resolve } from 'path';
 import { writeFile } from 'fs';
+import { resolve } from 'path';
 
 import * as TJS from 'typescript-json-schema';
-import { INLINES, BLOCKS } from '../src/index';
+import { BLOCKS, INLINES } from '../src/index';
 
 // optionally pass argument to schema generator
 const settings: TJS.PartialArgs = {
@@ -11,6 +11,7 @@ const settings: TJS.PartialArgs = {
   required: true,
   // @ts-ignore
   useTypeOfKeyword: true,
+  constAsEnum: true,
 };
 
 // optionally pass ts compiler options
@@ -25,7 +26,7 @@ const createJsonSchema = (symbolName: string, nodeType: string): void => {
 
   const schemaString = JSON.stringify(doc, null, 2);
 
-  writeFile(`./src/schemas/generated/${nodeType}.json`, schemaString, (err: Error) => {
+  writeFile(`./src/schemas/generated/${nodeType}.json`, schemaString, (err) => {
     if (err) {
       return console.log(err);
     }
@@ -53,6 +54,7 @@ const blockSymbolsMap = new Map([
   [BLOCKS.QUOTE, 'Quote'],
   [BLOCKS.EMBEDDED_ENTRY, 'EntryLinkBlock'],
   [BLOCKS.EMBEDDED_ASSET, 'AssetLinkBlock'],
+  [BLOCKS.EMBEDDED_RESOURCE, 'ResourceLinkBlock'],
   [BLOCKS.TABLE, 'Table'],
   [BLOCKS.TABLE_ROW, 'TableRow'],
   [BLOCKS.TABLE_CELL, 'TableCell'],
@@ -63,17 +65,19 @@ const inlineSymbolsMap = new Map([
   [INLINES.HYPERLINK, 'Hyperlink'],
   [INLINES.ENTRY_HYPERLINK, 'EntryHyperlink'],
   [INLINES.ASSET_HYPERLINK, 'AssetHyperlink'],
+  [INLINES.RESOURCE_HYPERLINK, 'ResourceHyperlink'],
   [INLINES.EMBEDDED_ENTRY, 'EntryLinkInline'],
+  [INLINES.EMBEDDED_RESOURCE, 'ResourceLinkInline'],
 ]);
 
 Object.values(BLOCKS).forEach((nodeType) => {
   const symbolName = blockSymbolsMap.get(nodeType);
-  createJsonSchema(symbolName, nodeType);
+  createJsonSchema(symbolName as string, nodeType);
 });
 
 Object.values(INLINES).forEach((nodeType) => {
   const symbolName = inlineSymbolsMap.get(nodeType);
-  createJsonSchema(symbolName, nodeType);
+  createJsonSchema(symbolName as string, nodeType);
 });
 
 createJsonSchema('Text', 'text');