+++ packages/rich-text-from-markdown/CHANGELOG.md
@@ -3,6 +3,80 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [15.17.4](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.17.3...@contentful/rich-text-from-markdown@15.17.4) (2024-06-25)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.17.3](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.17.2...@contentful/rich-text-from-markdown@15.17.3) (2024-05-28)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.17.2](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.17.1...@contentful/rich-text-from-markdown@15.17.2) (2024-05-27)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.17.1](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.17.0...@contentful/rich-text-from-markdown@15.17.1) (2024-05-24)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+# [15.17.0](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.15...@contentful/rich-text-from-markdown@15.17.0) (2024-05-22)
+
+### Features
+
+- upgrading rollup to latest version [TOL-2097] ([#559](https://github.com/contentful/rich-text/issues/559)) ([f14d197](https://github.com/contentful/rich-text/commit/f14d1974590d58f92bbf4cb56644095dba929ad9))
+
+## [15.16.15](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.14...@contentful/rich-text-from-markdown@15.16.15) (2024-03-04)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.14](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.13...@contentful/rich-text-from-markdown@15.16.14) (2024-03-01)
+
+### Bug Fixes
+
+- support <br/> from markdown ([#536](https://github.com/contentful/rich-text/issues/536)) ([30a3668](https://github.com/contentful/rich-text/commit/30a36685f42c70c98ae4bfb514ba710812ec0824))
+
+## [15.16.13](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.12...@contentful/rich-text-from-markdown@15.16.13) (2024-01-30)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.12](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.11...@contentful/rich-text-from-markdown@15.16.12) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.11](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.10...@contentful/rich-text-from-markdown@15.16.11) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.10](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.9...@contentful/rich-text-from-markdown@15.16.10) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.9](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.8...@contentful/rich-text-from-markdown@15.16.9) (2024-01-22)
+
+### Bug Fixes
+
+- [] Wrap inline nodes inside of a table cell in paragraph node ([#520](https://github.com/contentful/rich-text/issues/520)) ([6290f7d](https://github.com/contentful/rich-text/commit/6290f7dcae96c279d97e88ad65c42a080778fe73))
+
+## [15.16.8](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.7...@contentful/rich-text-from-markdown@15.16.8) (2023-09-12)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.7](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.6...@contentful/rich-text-from-markdown@15.16.7) (2023-08-04)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.6](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.5...@contentful/rich-text-from-markdown@15.16.6) (2023-05-26)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.5](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.4...@contentful/rich-text-from-markdown@15.16.5) (2023-05-04)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [15.16.4](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.3...@contentful/rich-text-from-markdown@15.16.4) (2023-03-14)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
 ## [15.16.3](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.16.2...@contentful/rich-text-from-markdown@15.16.3) (2023-02-09)
 
 ### Bug Fixes